Extract gitzen.models.github_{commit,pull_request}
commit-id:60890820
